+++ CHANGELOG.md
@@ -10,9 +10,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Bareos 25 and later will not modify or extend existing configuration directories on package updates [PR #2338]  
   Note: on FreeBSD unmodified configuration changes from Bareos <= 24 will get removed and
   replaced by the new default configuration of the current package.
+- Glusterfs FD plugin and SD backend is deprecated and will get removed in 26.
 
 ### Removed
 - config: deprecate file daemon as alias for client in FD config [PR #2187]
+- glusterfs: deprecate plugin and backend [PR #2416]
 
 ### Fixed
 - cats: fix version.map.in [PR #2064]
@@ -253,4 +255,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [PR #2403]: https://github.com/bareos/bareos/pull/2403
 [PR #2404]: https://github.com/bareos/bareos/pull/2404
 [PR #2412]: https://github.com/bareos/bareos/pull/2412
+[PR #2416]: https://github.com/bareos/bareos/pull/2416
 [unreleased]: https://github.com/bareos/bareos/tree/master